requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
diskcache==5.6.3
docker==6.1.3
FLAML==2.1.1
frozenlist==1.4.0
idna==3.4
install==1.3.5
multidict==6.0.4
numpy==1.26.1
openai==0.28.1
packaging==23.2
pyautogen==0.1.14
python-dotenv==1.0.0
requests==2.31.0
termcolor==2.3.0
tqdm==4.66.1
urllib3==2.0.7
websocket-client==1.6.4
yarl==1.9.2
